Add files via upload
A Bank wants to retain their existing customers who may churn, for that they have to know which existing customers has the highest probability of leaving the company, Built a Machine Learning
model which predicts the churning rate of customers in a particular bank, Used Logistic Regression, SVM, KNN and Random Forest to come up with the predictions. Got 89% accuracy on Random Forest Model.